chore: rename TooWideReturnTypeRector to NarrowTooWideReturnTypeRector
