--- rules-tests/DeadCode/Rector/FunctionLike/TooWideReturnTypeRector/config/configured_rule.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\DeadCode\Rector\FunctionLike\TooWideReturnTypeRector;
-
-return RectorConfig::configure()
-    ->withRules([TooWideReturnTypeRector::class]);
